use descriptive labels for nodes in tree view
Currently there is no way to provide a descriptive name for a node in the tree
view. You are stuck just using the name of the variable. This small patch
uses the label attribute from the node if available instead of the name to
allow finders to report a descriptive label for a node.

In our use case this allows us to include the interface description for a
router interface rather than just the interface name. This gives the user more
context to figure out what they are looking at.